refactor: 826 - move all constants from const index.ts to shared constants.ts
