+++ dev-data/courses.data.ts
@@ -1,4 +1,3 @@
-import { ROUTES } from '@/core/const';
 import type { Course } from '@/entities/course';
 import angular from '@/shared/assets/icons/angular.svg';
 import awsBlack from '@/shared/assets/icons/aws-black.svg';
@@ -8,7 +7,7 @@ import javascriptBlack from '@/shared/assets/icons/javascript-black.svg';
 import javascript from '@/shared/assets/icons/javascript.svg';
 import nodejs from '@/shared/assets/icons/node-js.svg';
 import react from '@/shared/assets/icons/react.svg';
-import { COURSE_LINKS, TO_BE_DETERMINED } from '@/shared/constants';
+import { COURSE_LINKS, ROUTES, TO_BE_DETERMINED } from '@/shared/constants';
 import { COURSE_TITLES } from 'data';
 
 export const courses: Course[] = [
