+++ dev-data/mentors-register.data.ts
@@ -1,4 +1,4 @@
-import { LINKS } from '@/core/const';
+import { LINKS } from '@/shared/constants';
 
 export const mentorsRegisterData = {
   en: {
