+++ src/core/base-layout/base-layout.test.tsx
@@ -2,7 +2,7 @@ import { render, screen } from '@testing-library/react';
 import { beforeEach, describe, expect, it, vi } from 'vitest';
 
 import { BaseLayout } from './base-layout';
-import { ROUTES } from '@/core/const';
+import { ROUTES } from '@/shared/constants';
 
 const mockUsePathname = vi.fn();
