+++ src/shared/__tests__/utils/render-with-router/render-with-router.tsx
@@ -3,7 +3,7 @@ import { render } from '@testing-library/react';
 import { RouterContext } from 'next/dist/shared/lib/router-context.shared-runtime';
 import { NextRouter } from 'next/router';
 
-import { ROUTES } from '@/core/const';
+import { ROUTES } from '@/shared/constants';
 
 interface RenderWithRouterProps {
   route?: string;
