+++ src/shared/__tests__/visual/not-found.spec.ts
@@ -1,7 +1,7 @@
 import { expect, test } from '@playwright/test';
 
 import { takeScreenshot } from './utils';
-import { ROUTES } from '@/core/const';
+import { ROUTES } from '@/shared/constants';
 
 test('Not-found page', async ({ page }) => {
   await page.goto(ROUTES.NOT_FOUND);
