+++ src/views/javascript-en.tsx
@@ -1,4 +1,4 @@
-import { ROUTES } from '@/core/const';
+import { ROUTES } from '@/shared/constants';
 import { getCourseLanguage } from '@/shared/helpers/get-course-language';
 import { AboutCourse } from '@/widgets/about-course';
 import { AboutVideo } from '@/widgets/about-video';
