+++ src/widgets/breadcrumbs/ui/breadcrumbs.tsx
@@ -5,7 +5,7 @@ import { usePathname } from 'next/navigation';
 
 import { BreadcrumbItem } from './breadcrumb-item';
 import { breadcrumbNameMap } from '../constants';
-import { ROUTES } from '@/core/const';
+import { ROUTES } from '@/shared/constants';
 import { RouteValues } from '@/shared/types';
 
 import styles from './breadcrumbs.module.scss';
