+++ src/widgets/upcoming-courses/ui/upcoming-courses.tsx
@@ -1,11 +1,11 @@
 import classNames from 'classnames/bind';
 import Image from 'next/image';
 
-import { ROUTES } from '@/core/const';
 import type { Course } from '@/entities/course';
 import { getCourses } from '@/entities/course/api/course-api';
 import { CourseItem } from '@/entities/course/ui/course-item/course-item';
 import RSBanner from '@/shared/assets/svg/RsBanner.svg';
+import { ROUTES } from '@/shared/constants';
 import { getActualData } from '@/shared/helpers/getActualData';
 import { LinkCustom } from '@/shared/ui/link-custom';
 import { Paragraph } from '@/shared/ui/paragraph';
